Improve Centos when source name is different from binary
This uses a fallback search when the main search fails and dives into
the repo metadata instead, using the `repomd` third party library.
This is able to ascertain the source rpm name given a binary.
This method ONLY works for v8 of Centos as the vault doesn't keep the
binary repodata around for older releases.

Other notable changes:
 - Changed the top-level vault dir to one that only has active
   releases.
 - Repo metadata downloads are cached so it should be quicker after the
   first call

https://jira-eng-sjc11.cisco.com/jira/browse/HELIOS-861