Extend caching behavior to all finders, add functional tests
Promote the get_url / test_url methods from the YumFinder to the
main SourceFinder, and refactor all of the finders to make use of
the common (cached) functionality.

Adds some token tests for all finders to demonstrate that they still
work as expected with the new caching layer.  The tests are disabled
by default.

Also updates the docstring for `SourceFinder.__init__` to explain
why caching is provided.

Turn the module-level *_TIMEOUT values into a common instance
property, so that it may be overridden, and add a `--timeout` switch
to the CLI to make use of it.