Partial revert "Extend caching behavior to all finders"
launchpadlib is not thread-safe, and its objects cannot be pickled
(https://bugs.launchpad.net/launchpadlib/+bug/822847) which means
we cannot use dogpile.cache with it, as it assumes both.

But the call to initialize launchpadlib is expensive, so attempt
to minimize remote calls within threads, at least.  This will cause
duplicated work in concurrent applications, which sadly cannot be
avoided.