Catch all failures from repomd loads, add warnings, docs
There are roughly eleven quintillion failure modes when loading
repomd, with zero of them being actionable in any realistic way.

Swallow all failures (not just HTTPErrors) and send back NoneTypes
from the subprocess, the callers will similarly roll those over
into "no package for you" results.

Since the YumFinder is designed with the full expectation that a
viable, persistent cache is in-place for it to do its work, add an
explicit warning if a YumFinder is initialized with the null cache

Add documentation on how to use caching.