Catch all failures from repomd loads, add warnings, docs (#22)
* Partial revert "Extend caching behavior to all finders"

launchpadlib is not thread-safe, and its objects cannot be pickled
(https://bugs.launchpad.net/launchpadlib/+bug/822847) which means
we cannot use dogpile.cache with it, as it assumes both.

But the call to initialize launchpadlib is expensive, so attempt
to minimize remote calls within threads, at least.  This will cause
duplicated work in concurrent applications, which sadly cannot be
avoided.

* Catch all failures from repomd loads, add warnings, docs

There are roughly eleven quintillion failure modes when loading
repomd, with zero of them being actionable in any realistic way.

Swallow all failures (not just HTTPErrors) and send back NoneTypes
from the subprocess, the callers will similarly roll those over
into "no package for you" results.

Since the YumFinder is designed with the full expectation that a
viable, persistent cache is in-place for it to do its work, add an
explicit warning if a YumFinder is initialized with the null cache

Add documentation on how to use caching.

* Re-raise rather than swallow exceptions, small cleanups

* Additional cleanups

Add an explanation for the `maxsize=` kwarg, since Python 3.7 still
needs it.

Make `TestYumFinderHelpers.test_do_task_reraises_exceptions` actually
run a job via do_task and ensure that the resulting exception is
intact.

* use exception chaining for additional clarity