Refactor the YumFinder (#38)
* Refactor the YumFinder

Implement our own XML fetch/parse routines for consuming repomd
data, and create a simple lookup dict from the results rather than
trying to shoehorn in use of the `repomd.findall` library call.

This yields a 30%-50% speedup in the functional tests.

* Drop yum.lookup_in_repomd function

Stepping through the refactored workflow showed that the YumFinder was
deep-copying a dict into a subprocess, looking up a value, and
returning the result.  Surely, there must be a better way.

Co-Authored-By: Nicolas Simonds <nisimond@cisco.com>